redirect-bridge/requirements.txt
azure-functions==1.20.0